prevent media picker from showing when capabilities.upload_files = 0
Problem: when in the “Contributor” role, the capability upload_files is
set to 0. In the editor, we’re still able to pick images, and we get
told the the upload failed.

Fix: store the capability in the Blog entity. When creating a post, if
the user taps the “add image” button, we check if they have the
capability or not. If they don’t, we show a tooltip explaining that
they can’t add images to their posts.